"Provide another group profile option: ""[  ] Full""
When a group is ""Full"" then:
1) Don't show it on any family's ""Available"" list
2) Do continue to show it elsewhere (e.g. do show it on a family's ""Confirmed list"", and do show it on the Staff: Groups page)"
Add a column for "Comments", showing the comments made by the family when they click to "Apply" or "Update" the group
"When creating a new account, it should have the status ""New Application"".

It does show as ""New Application"" on the Office Only page, however it's not showing the status on family search page until after clicking ""Save"" on the Office Only page"
Some of the tags show high # (e.g. 150) but don't show that many results
"Add these columns:
1) Potential (groups on their ""Available"" list)
2) Parent 2

Reword ""First Name"" to ""Parent 1"""
"The filters aren't functioning with Safari (HTML5 problem). Resolve this, e.g.
1) https://www.choicehomestay.com/family/list"
"The date fields aren't functioning with Safari(HTML5 problem) Resolve this, especially for:
1) Family Profile -> Profile
2) Family Profile -> Unavailability"